Linux fixes: Prevent Mono from throwing an ArugmentOutOfRangeException when the response is empty.
System.Web.HttpResponseStream.Write() throws an ArugmentOutOfRangeException for the offset argument under Mono when the passed Byte[] array
was created from an empty MemoryStream. (E.g. response.Length == 0) Even if the given offset is zero.

For example: This can happen when pkmn-classic-framework's GTS implementation has no data to return for a search.

So fix it by avoiding the Write() call when the response buffer is empty.